This change is about the refactoring of the source code of the objects use int the msg synchronization mechanism :
  - the functions xbt_os_mutex_lock() and xbt_os_mutex_unlock() are renamed xbt_os_mutex_acquire() and xbt_os_mutex_release() respectively
  - the functions xbt_os_sem_wait() and xbt_os_sem_post() are renamed xbt_os_sem_acquire() and xbt_os_sem_release() respectively
This renaming impacts the folowing files :

  - xbt_os_thread.h and xbt_os_thread.c
  - context.c
  - xbt_rl_synchro.c
  - run_context.c

In addition the function unschedule() declared an defined in the file context.c is renamed yield().

The prototype of the function xbt_os_sem_init() takes now only one parameter : the initial value of the semaphore. 
The second parameter of the function xbt_os_sem_timedacquire() which replace the function xbt_os_sem_timedwait() is now a double (desired timeout in seconds).
To finish, the semaphore interface of the xbt portability layer does not expose the functions xbt_os_sem_open() and xbt_sem_close().




git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@4848 48e7efb5-ca39-0410-a469-dd3cf9ba447f