This change introduce the new mailbox concept. A mailbox is a storage used to put or get an msg task. It is identified by an alias. For more information about the usage of the mailbox see the java examples alias0 and alias1 and the C example masterslave_forwarder_with_alias in the directory /examples/msg/alias.
This change impacts the following files :

include/msg/msg.h		Add the declaration of the functions related with the new concept of mailboxes.
src/msg/host.c			Adapte the functions __MSG_host_create and __MSG_host_destroy to the new concept of mailboxes.
src/gos/gos.c			Implementation of the functions related withe the new concept of mailboxes.
						
src/simix/smx_global.c	Initialization et finalization of the mailbox module.

src/java/jmsg.c			Implementation of the jni callback functions related with the mailbox concept.
src/java/jmsg.h			Declaration of the jni callback functions related with the mailbox concept.


ApplicationHandler.java	Implementation of the operations related with the properties of a process
Host.java				Implementation of the operations related with the new concept of mailbox
MsgNative.java			Implementation of the operations related with the new concept of mailbox
Process.java			Implementation of the operations related with the new concept of mailbox
Task.java				Implementation of the operations related with the new concept of mailbox


msg_mailbox.h			Header containing the declaration of the functions related with the mailbox concept
msg_mailbox.c			Source file containing the implementation of the functions related with the mailbox concept
		


directories alia0 et alias1 contain two examples to illustrate the different usages of the mailbox concept using the java interface
for msg.

directory alias contains the example masterslave_forwarder_with_alias which illustrate the usage of mailbox instead channel as in
the masterslave_forwarder example.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@5163 48e7efb5-ca39-0410-a469-dd3cf9ba447f