Speed up update_action_state AND generic_maxmin_share for sleep actions with no maximum duration.
These sleep actions are not put in the running_action swag but in
another private one. Thus, they're somehow hiden from the surf kernel
but it does not need to see them anyway. The cpu module takes care of
them as the only thing that might happen to them (from the surf kernel
perspective) would be to fail, which is now done through the
update_resource function.


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@5484 48e7efb5-ca39-0410-a469-dd3cf9ba447f