Gras listener thread of each process do select(-1) instead of while(1) {select(0.5)} So, it won't get active before a message actually arrived.
The previous design was because when the main thread opens a server
socket, it has to be added to the set of sockets select()ed by the
listener. My solution is to add a new socket between main and listener
threads. Each time the main opens a socket, it writes something on
this socket, unlocking the listener. Who detects the fact that it's
not a real message to handle, and get back to select(-1)
(automagically adding the new socket to the set).


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@6151 48e7efb5-ca39-0410-a469-dd3cf9ba447f