Various fixes. Working (in SG) for ForkJoins, but not for Montage. * Change synchronization scheme for ending: workers send go to master   instead of having the master calling amok_pm_group_end() right after   the setup. This killed the workers ways too early. * make sure that workers do create all mandatory sockets to others   since dopar threads have real difficulties speaking to the listener   thread to add new sockets * instead of having a callback to the go message in the worker which   does all the work, do explicitely wait for that message and move the   code in worker's main()
Like I said, there is still a deadlock when trying to replay
Montage_25, but this may related to the fact that this workflow seems
somehow broken. the "cimages.tbl" file, generated by root and consumed
by ID00021@mImgTbl is mangled in the dot output, and thus presumably
also in memory representation. More to come tomorrow.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@6792 48e7efb5-ca39-0410-a469-dd3cf9ba447f