Fight for better integration of mmalloc, mc and xbt
I wanted to move the legacy interception (ie, redefinition of
malloc/free/realloc) from mc to mmalloc, where it should be.

But it broke libgras, with memory curruption symptoms in tesh. I
thought that it was a race condition because mmalloc is not reentrent.

So, I protected the main functions with a xbt_os_mutex (and creating
it in the default heap was challenging since the mutex itself must be
in the heap, which is not ready yet).

But it didn't help: I still had the memory corruption symptoms in
tesh, inchanged. Maybe some system function get fucked up by the
mmalloc stuff?

So I decided to comment out the legacy interception from mmalloc for a
while (breaking mc, but who cares at this point?).

But the xbt_os_mutexes are not in libsimgrid since pthreads and
ucontextes didn't intermix well a few years ago. So, I changed the
corresponding stubs from "die loudly" to "do nothing". That's
dangerous (one could think he's protected by mutex when not), but I
need the EPR refactoring to sort it out correctly.

All this for... nothing. The legacy interception will still be in mc.
And I guess that mc must be in another library, not the simulation
one, if it does not intermix well with syscalls.

Well mmalloc is now a bit more thread-friendly, let's be positive.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@7711 48e7efb5-ca39-0410-a469-dd3cf9ba447f