Let it compile in paranoid mode without _XBT_GNUC_UNUSED (cosmetics only)
The main issue of the previous fix to the problem (by Arnaud G) was
that it documented the commit in the source, explaining why it was
mandatory in a C comment. That's source bloat in my mind:
git blame+git log can give you the same information if you need it.

So I figured out an excuse to recommit on top of it, and here we go :)

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@8512 48e7efb5-ca39-0410-a469-dd3cf9ba447f