nah, properties in the XML file are not the right place to pass configuration options
Actually, the properties are the right location, but only when
included in a config tag. And in this setting, the right behavior is
already done by the </config> handler, no need to also do it for the
generic <prop> handler. Doing so anyway would prevent to pass
"coordinates" (user-)properties to hosts, links and such.