Simplify the mmalloc library further
* Stop playing with void* for the heap descriptors and introduce the
  xbt_mheap_t datatype for that.
* Don't try to make it work when mmalloc and friends are called with
  NULL as a first argument.

I added a pimple that mmalloc_preinit returns the default mhead after
creating it to simplify the code of the legacy functions that check
that this default mhead exist or create it before using it.

This extra check should be useless if mmalloc_preinit were called soon
enough since it's in the critical path, but the performance of
model-checking is not a concern yet: only getting something working
matters for now.