Revert "Don't fail when backtrace() returns 0."
This reverts commit 944109bb21e5dc635a47813a2adff0e16305d9c9.

This commit is controversial, and was intended to fix build on systems
where backtrace exists, but returns 0.

The only known systems with such behavior are Debian on ARM (armel, armhf).
Another workaround is to compile with the -funwind-tables gcc option.

See http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=637360 for more information.