Platform generation can now be used with MSG
Some stuff had to be done by SIMIX and MSG right after the XML platform parsing. It is now done by callbacks, registered in SURF at initialization. So that stuff is also done when you create a random platform with the generator.
Add some ugly code to avoid the deletion of these registered callbacks before the parsing.