Do not kill all processes when MSG shuts down: simix will cleanup anyway
Plus, it broke Java since this function is now called using atexit(),
and it seems that in Java, the thread running these functions is not
maestro. So our mechanism was trying to fire a simcall from this
thread, which is a Java system thread that is unknown from SimGrid.

As a result, the JVM was segfaulting when exiting.