use MSG_init instead of xbt_init since someone wrongly create mutexes
There should not be necessary to call MSG_init to create xbt_mutexes,
but since this creation induces a simcall, we'd better initialize
simix.

This is not sufficient though, as maestro_context does not seem to be
initialized this way, leading to a try to dereference the host of the
current process, that was never set.