Remove comm from src_proc->comms and dst_proc->comms when finish it, if is detached (dsend)
Currently, with dsend action, the wait of src_proc is not necessary to
finish the communication. In this case, only dst_proc removes this
communication from its list and destroys it, if src_proc is always
running. When src_proc makes a process_cleanup after that, it
retrieves this communication in its list while it doesn't exist
anymore and thinks that this communication isn't for it. This led to a
simulation error.