add a model object to the arguments of callbacks
For VM support, we have multiple cpu model objects on the system. The
object of a surf model may not be only one on the system. Thus, we have
to modify the code that assumes the system has only one global object of
each surf model.

We add a model object to the first argument of the callbacks of struct
surf_model_private; i.e., share_resources(), update_actions_state(), and
finalize(). In each callbacks, we delete the use of the global model
object.