model-checker : last fix in DPOR algorithm if max depth is reached
If max_depth is reached, the last state in stack may have a process
interleaved with a request not executed yet. In that case, we start to
apply the independence theorem with this last request not executed
yet. Otherwise (no more process to interleave in the last state of the
stack), last state is deleted before starting to apply independence theorem.