Add convenient scripts to check spelling.
I stole them from http://www.kegel.com/kerspell/ and updated them to "modernity" (aspell did not work as expected so I switched to ispell).
Simply sh tools/spell/spell.sh and it will spit a list of dubious words.
Note that it ignores all the dubious words listed in tools/spell/sg_stopwords.txt and that I only checked the A's in there so far.