Add a test program to see multicore behaviors.
As far as tested through this test program, it looks that the current
support of multicore in SimGrid correctly works for physical and virtual
machines. The CPU share of a task or a VM never exceeds the capacity of
one of the CPU cores. The total CPU share of all tasks and VMs on a PM
never exceeds the product of the number of cores and the capacity of a core.
(But, it is not possible to pin a task to a specific core.)

TODO: use this program for regression test.