[mc] Fix bug when trying to handle DW_OP_regN in MC_dwarf_resolve_location
MC_dwarf_resolve_location is supposed to return the *address* of a
given variable. However, the DW_OP_regN returned the value of the
register and not an address address of a variable with its content.

It should be possible to get such an addresse (the registers are saved
on the stack) but the feature does not seem to be implemented in
libunwind.