Begin to add an interface to instanciate SMPI programatically.
This adds SMPI_init, SMPI_finalize, and SMPI_app_instance_register

This allows user to deploy MSG+MPI applications in the same simulation,
as well as deploying several instances of SMPI with the same or different application,
without using smpirun. This means that we can now handle several MPI_COMM_WORLD.

This is only a very early step, as for now the interface still depends on MSG calls
 to deploy the application, and run the processes.

We should work on an interface to do this for SMPI, or add a generic interface
that would allow several SimGrid interfaces to run alltogether.