Set 4 tests that fail on Windows and won't ever pass as "expected to fail"
This will allow to better discriminate between bugs and wontfix non-issues.

One of them fails because random returns differently on win and generates different colors in the resulting trace. Duh.
The three others are expected to crash, and do crash correctly on Windows, but the return code is different from the one expected (SIGABRT). Meh