Tesh: Use 0 as a default sorting prefix
I used 'output sort' to deal with events that occure in order of map
traversal, and could not devise why it was uneffective. Until I digged
deep in tesh again.

So use a sensible, unsurprising value, and change it where needed. And
update the documentation, just in case someone would read it in a
parallel dimension.