Channel: the S4U name of msg_mailbox_t
I hope that the name Channel will be easier to understand than mailbox
(users were often thinking that mailboxes must be attached to a given
host).

The communication primitives do not take a name to a mailbox, but a
mailbox directly. I hate python for raising no warning when I do a
typo on a variable name, and addressing mailboxes by name leads to the
exact same situation.

I would like Channel::byName to return a reference (since they are
created -- in kernel mode -- on need), but I fail to get the syntax
right.

At some point, a C API should be exported, the content of smx_rdv_t
should be integrated into that class, and smx_rdv_t should be killed.
But I want to get a working API to play with first.